scripts/models/requirements.txt
# Model conversion requirements
# Run: pip install -r requirements.txt

torch>=2.0.0
torchvision>=0.15.0
safetensors>=0.4.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0
requests>=2.31.0
scikit-learn>=1.3.0

# For eye state training
# Dataset: CEW (Closed Eyes in the Wild)
